requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
numpy==1.24.3
pyarrow==14.0.1
pillow
protobuf<4.0,>=3.20